Fix new controllers pairing/joining with bluez5
When a brand new controller is paired the bluetooth stack is restarted
by psmoveapi. In previous versions of bluez this wasn't a problem,
however in bluez5 (shipped in the latest raspbian) the bluetooth device
power state is not persisted across restarts - udev activates it on
boot, but if you /etc/init.d/bluetooth restart the adapter will not be
powered on.

This means that whenever you connect a brand new controller over USB to
pair it you have to cycle the rpi to get the bluetooth re-enabled.

This changes scan_enable to also power up the devices constantly,
working around the problem. Ugly, but whatever.